refactor(order): Refactor order management services and integrate inventory handling
- Replaced OrderStockService with OrderInventoryIntegrationService to centralize inventory management.
- Updated OrderManagementService to utilize the new OrderInventoryIntegrationService for stock validation and material consumption.
- Refactored OrderHistoryController, OrderStatusController, and other controllers to use OrderManagementService instead of the deprecated OrderService.
- Removed OrderService and its related tests, consolidating functionality into OrderManagementService.
- Adjusted dependency injection in provider wiring to reflect the new service structure.
- Updated tests to mock OrderManagementService instead of the deprecated OrderService.
